Add files via upload
SensorTag-for-Windows
=====================

Sample code that talks to Texus Instruments SensorTag Bluetooth LE device for Windows Universal Apps.  This code supports both versions CC2541 and CC2650

First pair your SensorTag with your Windows 8.1 device, then this app will automatically connect to it and show you all the data.

The SensorTagLibrary code is the part that is designed to be reused in your apps, it provides most of the bluetooth characteristic data available on the SensorTag.